+++ src/main/java/com/silverminer/shrines/utils/network/cts/CTSFetchNovelAmountPacket.java
@@ -1,15 +1,19 @@
-package com.silverminer.shrines.utils.network;
+package com.silverminer.shrines.utils.network.cts;
 
 import java.util.UUID;
 import java.util.function.Supplier;
 
 import com.silverminer.shrines.structures.load.StructureData;
 import com.silverminer.shrines.structures.novels.NovelsDataRegistry;
+import com.silverminer.shrines.utils.network.IPacket;
+import com.silverminer.shrines.utils.network.ShrinesPacketHandler;
+import com.silverminer.shrines.utils.network.stc.STCFetchNovelsAmountPacket;
 
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.server.MinecraftServer;
+import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.LogicalSide;
 import net.minecraftforge.fml.LogicalSidedProvider;
 import net.minecraftforge.fml.network.NetworkEvent;
@@ -38,11 +42,26 @@ public class CTSFetchNovelAmountPacket implements IPacket {
 
 	public static void handle(CTSFetchNovelAmountPacket packet, Supplier<NetworkEvent.Context> context) {
 		context.get().enqueueWork(() -> {
-			MinecraftServer server = LogicalSidedProvider.INSTANCE.get(LogicalSide.SERVER);
-			ShrinesPacketHandler.sendTo(
-					new STCFetchNovelsAmountPacket(packet.structure, NovelsDataRegistry.getNovelAmount(packet.structure.getKey())),
-					server.getPlayerList().getPlayer(packet.player));
+			Handle.handle(packet);
 		});
 		context.get().setPacketHandled(true);
+	}
+
+	private static class Handle {
+		public static DistExecutor.SafeRunnable handle(CTSFetchNovelAmountPacket packet) {
+			return new DistExecutor.SafeRunnable() {
+
+				private static final long serialVersionUID = 1L;
+
+				@Override
+				public void run() {
+					MinecraftServer server = LogicalSidedProvider.INSTANCE.get(LogicalSide.SERVER);
+					ShrinesPacketHandler.sendTo(
+							new STCFetchNovelsAmountPacket(packet.structure,
+									NovelsDataRegistry.getNovelAmount(packet.structure.getKey())),
+							server.getPlayerList().getPlayer(packet.player));
+				}
+			};
+		}
 	}
 }